requirements.txt
future
gevent
Pillow
matplotlib
nibabel
numpy
pandas
dicom
pydicom
pyparsing
scikit-learn
scipy
SimpleITK
six
tensorflow
sklearn
PyQt5
rx
zerorpc
skimage